# perf: use Observable.Create instead of FromEvent