I'll wait to update the benchmark detail files with a full set of benchmarks once I finish up with the source generator.

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.630 (2004/?/20H1)
Intel Core i5-4200U CPU 1.60GHz (Haswell), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=5.0.100
  [Host]        : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  .NET Core 3.1 : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT

Job=.NET Core 3.1  Runtime=.NET Core 3.1

|                  Method | Changes |         Mean |      Error |       StdDev |       Median | Ratio | RatioSD |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|------------------------ |-------- |-------------:|-----------:|-------------:|-------------:|------:|--------:|----------:|------:|------:|-----------:|
| Change_Depth1_FromEvent |    1000 |     66.21 us |   1.346 us |     3.949 us |     66.93 us |  1.00 |    0.00 |   15.2588 |     - |     - |   23.44 KB |
|    Change_Depth1_Create |    1000 |     67.52 us |   1.377 us |     4.059 us |     69.32 us |  1.02 |    0.09 |   15.2588 |     - |     - |   23.44 KB |
|                         |         |              |            |              |              |       |         |           |       |       |            |
| Change_Depth2_FromEvent |    1000 |  3,673.66 us |  72.890 us |   206.777 us |  3,716.61 us |  1.00 |    0.00 |  933.5938 |     - |     - | 1434.16 KB |
|    Change_Depth2_Create |    1000 |  2,040.23 us |  40.982 us |   120.838 us |  2,077.17 us |  0.56 |    0.05 |  470.7031 |     - |     - |  720.99 KB |
|                         |         |              |            |              |              |       |         |           |       |       |            |
| Change_Depth3_FromEvent |    1000 |  8,265.34 us | 209.591 us |   608.061 us |  8,562.31 us |  1.00 |    0.00 | 1843.7500 |     - |     - | 2841.08 KB |
|    Change_Depth3_Create |    1000 |  4,364.26 us |  98.897 us |   288.488 us |  4,479.16 us |  0.53 |    0.06 |  921.8750 |     - |     - | 1416.21 KB |
|                         |         |              |            |              |              |       |         |           |       |       |            |
| Change_Depth4_FromEvent |    1000 | 12,880.64 us | 378.402 us | 1,109.788 us | 13,121.44 us |  1.00 |    0.00 | 2765.6250 |     - |     - | 4253.37 KB |
|    Change_Depth4_Create |    1000 |  6,974.76 us | 123.850 us |   115.849 us |  6,984.71 us |  0.50 |    0.04 | 1375.0000 |     - |     - | 2114.13 KB |
```